+++ app/register/page.tsx
@@ -1,18 +1,30 @@
 "use client"
 
+import { PingAnimation } from "@/app/components/animation/pingAnimation"
 import { AlertContext } from "@/app/components/layout/alertBox"
 import { Stepper } from "@/app/components/stepper"
 import type { Alert } from "@/app/interfaces/alert"
 import type { FormItem } from "@/app/interfaces/formItem"
 import type { Profile } from "@/app/interfaces/profile"
 import { CHECKLIST } from "@/app/lib/constant"
-import { CheckIcon } from "@heroicons/react/24/solid"
+import {
+  ArrowUturnLeftIcon,
+  CheckIcon,
+  PaperAirplaneIcon,
+  XMarkIcon,
+} from "@heroicons/react/24/solid"
+import Image from "next/image"
 import { useRouter } from "next/navigation"
 import type React from "react"
 import { useContext, useRef } from "react"
-import { type Path, type UseFormRegister, useForm } from "react-hook-form"
-import { ConfirmDialog } from "./components/confirmDialog"
-import { ImageUploader } from "./components/imageUploader"
+import {
+  type Path,
+  type SubmitHandler,
+  type UseFormRegister,
+  type UseFormWatch,
+  useForm,
+} from "react-hook-form"
+import { ImageUploader } from "./imageUploader"
 
 const COMPANIES: string[] = [
   "オープンアップグループ",
@@ -33,7 +45,7 @@ export default function Register(): React.JSX.Element {
     useContext(AlertContext)
   const router = useRouter()
 
-  async function onSubmit(): Promise<void> {
+  const onSubmit: SubmitHandler<Profile> = async () => {
     const formElement: HTMLFormElement = formRef.current as HTMLFormElement
     const formData: FormData = new FormData(formElement)
 
@@ -42,8 +54,8 @@ export default function Register(): React.JSX.Element {
       body: formData,
     })
       .then((res) => {
-        dialogRef.current?.close()
         if (!res.ok) {
+          dialogRef.current?.close()
           setAlert({ eventType: "error", message: res.statusText })
           return
         }
@@ -170,5 +182,82 @@ function Input({
         required={true}
       />
     </label>
+  )
+}
+
+function ConfirmDialog({
+  ref,
+  watch,
+  isSubmitting,
+}: Readonly<{
+  ref: React.RefObject<HTMLDialogElement>
+  watch: UseFormWatch<Profile>
+  isSubmitting: boolean
+}>): React.JSX.Element {
+  return (
+    <dialog ref={ref} className="modal modal-bottom sm:modal-middle">
+      <div className="grid gap-4 modal-box text-center">
+        <button
+          type="button"
+          onClick={() => ref.current?.close()}
+          className="btn btn-sm btn-circle btn-ghost absolute right-2 top-2 hover:scale-110"
+          aria-label="閉じる"
+        >
+          <XMarkIcon />
+        </button>
+        <Stepper targetStep={1} />
+        <table className="table">
+          <tbody>
+            {CHECKLIST.map(({ name, value }) => (
+              <tr key={name}>
+                <th>{value}</th>
+                <td>
+                  {name === "agreement" && "同意する"}
+                  {name === "image" && watch(name as Path<Profile>) ? (
+                    <Image
+                      src={document.getElementsByTagName("img")[0].src}
+                      width={100}
+                      height={100}
+                      alt="Uploaded File"
+                      className="w-full"
+                    />
+                  ) : (
+                    (watch(name as Path<Profile>) as string)
+                  )}
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+        <div className="modal-action justify-center gap-4">
+          <button
+            type="submit"
+            className={`btn btn-info ${
+              isSubmitting ? "indicator" : "[&:not(:hover)]:animate-bounce"
+            }`}
+            disabled={isSubmitting}
+          >
+            {isSubmitting && <PingAnimation />}
+            <PaperAirplaneIcon className="size-6" />
+            送信
+          </button>
+          <button
+            type="button"
+            className="btn btn-error hover:scale-110"
+            onClick={() => ref.current?.close()}
+          >
+            <ArrowUturnLeftIcon className="size-6" />
+            戻る
+          </button>
+        </div>
+      </div>
+      <div className="modal-backdrop">
+        <button
+          type="button"
+          onClick={() => ref.current?.close()}
+          aria-label="戻る"
+        />
+      </div>
+    </dialog>
   )
 }
